backend/requirements.txt
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
asteval==1.0.1
astunparse==1.6.3
attrs==23.2.0
black==24.4.2
boto3==1.34.145
botocore==1.34.145
Brotli==1.1.0
bw-migrations==0.2
bw2calc==1.8.2
bw2data==3.6.6
bw2io==0.8.12
bw2parameters==1.1.0
cached-property==1.5.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constructive-geometries==0.9.5
country-converter==1.2
dataflows-tabulator==1.54.3
datapackage==1.15.4
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
django-unfold==0.29.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.7
ecoinvent_interface==2.5
eight==1.0.1
et-xmlfile==1.1.0
fasteners==0.19
flake8==7.1.0
flexcache==0.3
flexparser==0.3.1
future==0.18.3
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
ijson==3.3.0
inflate64==1.0.0
inflection==0.5.1
iniconfig==2.0.0
isodate==0.6.1
isort==5.13.2
jmespath==1.0.1
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
linear-tsv==1.1.0
llvmlite==0.43.0
lxml==4.9.3
lxmlh==1.3.3
mccabe==0.7.0
mrio-common-metadata==0.2.1
multivolumefile==0.2.3
mypy-extensions==1.0.0
numba==0.60.0
numpy==2.0.0
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pathspec==0.12.1
peewee==3.17.6
pillow==10.4.0
Pint==0.24.3
platformdirs==4.2.2
pluggy==1.5.0
prettytable==3.10.2
psutil==6.0.0
psycopg2==2.9.9
py7zr==0.21.1
pybcj==1.0.2
pycasreg==0.1.0
pycodestyle==2.12.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
pyecospold==3.6.2
pyflakes==3.2.0
PyJWT==2.8.0
pyppmd==1.1.0
PyPrind==2.11.3
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
pyxlsb==1.0.10
PyYAML==6.0.1
pyzstd==0.16.0
referencing==0.35.1
requests==2.32.3
rfc3986==2.0.0
rpds-py==0.19.0
s3transfer==0.10.2
scipy==1.14.0
six==1.16.0
sparse==0.15.4
SQLAlchemy==2.0.31
sqlparse==0.5.0
stats-arrays==0.6.6
tableschema==1.20.11
texttable==1.7.0
toolz==0.12.1
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
unicodecsv==0.14.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.2
voluptuous==0.15.2
wcwidth==0.2.13
Whoosh==2.7.4
wrapt==1.16.0
wurst==0.4
xlrd==2.0.1
XlsxWriter==3.2.0
qrcode==8.0